Move creating doc mention select menu inside controller
